Use pip download

+++ pydep-run.py
@@ -6,18 +6,21 @@ import argparse
 import pydep.req
 import pydep.setup_py
 
-from os import system, path
+from os import path
 import subprocess
 import tempfile
 import shutil
 
+
 def main():
     # Parse args
-    argparser = argparse.ArgumentParser(description='pydep is simple command line tool that tells you about package dependency metadata in Python')
+    argparser = argparse.ArgumentParser(
+        description='pydep is simple command line tool that tells you about package dependency metadata in Python')
     subparsers = argparser.add_subparsers()
 
     dep_parser = subparsers.add_parser('dep', help='print the dependencies of a python project in JSON')
-    dep_parser.add_argument('--raw', action='store_true', help='If true, pydep will not try to resolve dependencies to VCS URLs')
+    dep_parser.add_argument('--raw', action='store_true',
+                            help='If true, pydep will not try to resolve dependencies to VCS URLs')
     dep_parser.add_argument('dir', help='path to root directory of project code')
     dep_parser.set_defaults(func=dep)
 
@@ -29,7 +32,9 @@ def main():
     list_info_parser.add_argument('dir', help='path to containing directory')
     list_info_parser.set_defaults(func=list_info)
 
-    smoke_parser = subparsers.add_parser('demo', help='run pydep against some popular repositories, printing out dependency information about each')
+    smoke_parser = subparsers.add_parser('demo',
+                                         help='run pydep against some popular repositories, printing out dependency'
+                                              ' information about each')
     smoke_parser.set_defaults(func=smoke_test)
 
     args = argparser.parse_args()
@@ -39,6 +44,7 @@ def main():
 #
 # Sub-commands
 #
+
 
 def list_info(args):
     """Subcommand to print out metadata of all packages contained in a directory"""
@@ -50,8 +56,10 @@ def list_info(args):
         if err is not None:
             sys.stderr.write('failed due to error: %s\n' % err)
             sys.exit(1)
-        setup_infos.append(setup_dict_to_json_serializable_dict(setup_dict, rootdir=path.relpath(setup_dir, container_dir)))
-    print json.dumps(setup_infos)
+        setup_infos.append(
+            setup_dict_to_json_serializable_dict(setup_dict, rootdir=path.relpath(setup_dir, container_dir)))
+    print(json.dumps(setup_infos))
+
 
 def info(args):
     """Subcommand to print out metadata of package"""
@@ -59,7 +67,8 @@ def info(args):
     if err is not None:
         sys.stderr.write('failed due to error: %s\n' % err)
         sys.exit(1)
-    print json.dumps(setup_dict_to_json_serializable_dict(setup_dict))
+    print(json.dumps(setup_dict_to_json_serializable_dict(setup_dict)))
+
 
 def dep(args):
     """Subcommand to print out dependencies of project"""
@@ -67,7 +76,8 @@ def dep(args):
     if err is not None:
         sys.stderr.write('failed due to error: %s\n' % err)
         sys.exit(1)
-    print json.dumps(reqs)
+    print(json.dumps(reqs))
+
 
 def smoke_test(args):
     """Test subcommand that runs pydep on a few popular repositories and prints the results."""
@@ -77,42 +87,45 @@ def smoke_test(args):
         # TODO: update smoke_test to call setup_dirs/list instead of assuming setup.py exists at the repository root
         # ('Node', 'https://github.com/joyent/node.git'),
     ]
+    tmpdir = None
     try:
         tmpdir = tempfile.mkdtemp()
         for title, cloneURL in testcases:
-            print 'Downloading and processing %s...' % title
+            print('Downloading and processing %s...' % title)
             subdir = path.splitext(path.basename(cloneURL))[0]
             dir_ = path.join(tmpdir, subdir)
             with open('/dev/null', 'w') as devnull:
                 subprocess.call(['git', 'clone', cloneURL, dir_], stdout=devnull, stderr=devnull)
 
-            print ''
+            print('')
             reqs, err = pydep.req.requirements(dir_, True)
             if err is None:
-                print 'Here is some info about the dependencies of %s' % title
+                print('Here is some info about the dependencies of %s' % title)
                 if len(reqs) == 0:
-                    print '(There were no dependencies found for %s)' % title
+                    print('(There were no dependencies found for %s)' % title)
                 else:
-                    print json.dumps(reqs, indent=2)
+                    print(json.dumps(reqs, indent=2))
             else:
-                print 'failed with error: %s' % err
+                print('failed with error: %s' % err)
 
-            print ''
+            print('')
             setup_dict, err = pydep.setup_py.setup_info_dir(dir_)
             if err is None:
-                print 'Here is the metadata for %s' % title
-                print json.dumps(setup_dict_to_json_serializable_dict(setup_dict), indent=2)
+                print('Here is the metadata for %s' % title)
+                print(json.dumps(setup_dict_to_json_serializable_dict(setup_dict), indent=2))
             else:
-                print 'failed with error: %s' % err
+                print('failed with error: %s' % err)
 
     except Exception as e:
-        print 'failed with exception %s' % str(e)
+        print('failed with exception %s' % str(e))
     finally:
-        shutil.rmtree(tmpdir)
+        if tmpdir:
+            shutil.rmtree(tmpdir)
 
 #
 # Helpers
 #
+
 
 def setup_dict_to_json_serializable_dict(d, **kw):
     return {
